requirements.txt
DAWG-Python==0.7.1
Django==1.7.1
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.7.0
PyYAML==3.11
Pygments==2.0.1
Tempita==0.5.2
Unidecode==0.04.17
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
billiard==3.3.0.19
celery==3.1.17
certifi==14.05.14
clepy==0.3.21
decorator==3.4.0
django-appconf==1.0.1
django-bootstrap-breadcrumbs==0.7.0
django-celery==3.1.16
django-compressor==1.4
django-custom-user==0.5
django-debug-toolbar==1.2.2
django-easymode==1.4b5
django-filebrowser==3.5.7
-e git://github.com/smacker/django-filebrowser-no-grappelli.git@911318f0e735ab60485c5783759b3f6a058f2337#egg=django_filebrowser_no_grappelli-master
django-grappelli==2.5.6
django-haystack==2.3.1
django-htmlmin==0.7.0
django-memcache-status==1.1
django-postman==3.2.0
django-reversion==1.5.1
django-robots==1.0
django-rosetta==0.7.4
django-simple-captcha==0.4.4
django-tinymce==1.5.3
docopt==0.6.2
docutils==0.12
easy-thumbnails==2.2
elasticsearch==1.2.0
flower==0.7.3
html5lib==0.999
kombu==3.0.24
mechanize==0.2.5
mock==1.0.1
nose==1.3.4
polib==1.0.4
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
python-dateutil==2.4.0
python-memcached==1.53
pytz==2014.10
redis==2.10.3
requests==2.5.0
six==1.8.0
sqlparse==0.1.14
tornado==4.0.2
unicode-slugify==0.1.3
urllib3==1.9.1
version==0.1.0
wsgiref==0.1.2
